Graduation_project/smart_doc_qua/requirements.txt
fastapi
uvicorn
pypdf2
langchain
faiss-cpu
sentence-transformers
numpy
python-multipart
requests
openai>=1.0